+++ src/state/posts/atoms.ts
@@ -1,5 +1,5 @@
 import { atom } from 'recoil';
-import type { TPost } from '@/api/index';
+import type { TPost } from '@/requests/index';
 
 export const postsState = atom({
   key: 'postsState',
