+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
-      "@/api/*": ["api/*"],
+      "@/requests/*": ["requests/*"],
       "@/components/*": ["components/*"],
       "@/common-components": ["components/common/*"],
       "@/page-components": ["components/page/*"],
